+++ packages/components/src/Combobox/ComboboxProvider.tsx
@@ -1,17 +1,6 @@
 import React, { Dispatch, MutableRefObject, SetStateAction } from "react";
 import { ComboboxOption } from "./Combobox.types";
 
-export const ComboboxContext = React.createContext(
-  {} as {
-    open: boolean;
-    setOpen: (open: boolean) => void;
-    selected: ComboboxOption[];
-    selectionHandler: (option: ComboboxOption) => void;
-    handleClose: () => void;
-    shouldScroll: MutableRefObject<boolean>;
-  },
-);
-
 export interface ComboboxProviderProps {
   readonly children: React.ReactNode;
   readonly selected: ComboboxOption[];
@@ -20,23 +9,20 @@ export interface ComboboxProviderProps {
   readonly setOpen: Dispatch<SetStateAction<boolean>>;
   readonly handleClose: () => void;
   readonly shouldScroll: MutableRefObject<boolean>;
+  readonly searchValue: string;
 }
 
-export function ComboboxContextProvider(
-  props: ComboboxProviderProps,
-): JSX.Element {
+export const ComboboxContext = React.createContext(
+  {} as Omit<ComboboxProviderProps, "children">,
+);
+
+export function ComboboxContextProvider({
+  children,
+  ...props
+}: ComboboxProviderProps): JSX.Element {
   return (
-    <ComboboxContext.Provider
-      value={{
-        open: props.open,
-        setOpen: props.setOpen,
-        selected: props.selected,
-        selectionHandler: props.selectionHandler,
-        handleClose: props.handleClose,
-        shouldScroll: props.shouldScroll,
-      }}
-    >
-      {props.children}
+    <ComboboxContext.Provider value={props}>
+      {children}
     </ComboboxContext.Provider>
   );
 }